Page Update [61] | Voice Chat Update [2] | Add Message\components VoiceCallModal
Within the scope of UI/UX Design Refresh (Phase 6) and Voice Chat Technology Implementation:
Page Update [61] and Voice Chat Update [2]: have been completed. I updated the design of the voice chat component on the communication page and added a sidebar named VoiceCallSidebar to provide quick access to the voice chat screen. Then, I added a range input to adjust the voice detection level with the voice chat settings button on the sidebar. Testing for voice chat is still ongoing. It can be tested to see how the design functions.