Page Update [74]
Within the scope of the UI/UX Design Overhaul (Phase 8), I implemented Page Update [74]. We made enhancements to the user interface design to improve aesthetics. With these changes, previously we directed post content to the PostDetails page using Routes in app.tsx. Now, we have updated the necessary functionality for users to be directed upon clicking.