Part Update [40]
Within the scope of the UI/UX Design Overhaul (Phase 8), I conducted Part Update [40]. I updated the profile panel for users on the profile page. Within the ProfileHeaderProps, I connected the navigation section to the parent page, which is the profile page, using props. This way, we established a dynamic content view structure.