Part Update [52]
Within the scope of the UI/UX Design Overhaul (Phase 8), Part Update [52] implemented. Within this update, I refined the design of the ProfileFriendsContent component under the Profile page, enhancing its visual appeal. Additionally, it was connected to the Context API. This allows both the user and visitors to the profile to see the user's friends according to the restrictions set by the user.