Part Update [55] (Video Content Support Added)
Within the scope of the UI/UX Design Overhaul (Phase 8), Part Update [55] has been implemented. As part of this update, I enhanced the design of the ProfileMainContent component, improving its visual appeal. Now, there are different designs for users based on their post types. If there is a photo in the post, I also added a popup for users to view the photo in detail. Video Content Support Added.