Page Update [94] (MessagePage)
Within the scope of the UI/UX Design Overhaul (Phase 8), Page Update [94] has been implemented. As part of this update, I connected the messaging interface between users on the MessagePage to the API. Mina Protocol transactions are ongoing. Additionally, I created a portable smart bar, which is an important step towards creating a system that will provide ease of use in the future of the project.