Page Update [95] (MyGroups)
Within the scope of the UI/UX Design Overhaul (Phase 8), Page Update [95] has been implemented. As part of this update, I added necessary redirection in case of a network error when the API connection cannot be established on the MyGroups page. Additionally, I ensured that only authorized and authenticated users can fetch data from the API by using Authorization Bearer in the design and API communication.