Page Update [97] (MessageRequests)
Within the scope of the UI/UX Design Overhaul (Phase 8), Page Update [97] has been implemented. In this update, I made design enhancements to the MessageRequests component on the CommunicationPage and implemented the functionality to list the message requests for the user. Only authorized and authenticated users can view their own message requests.