Page Update [98] (GroupInvitation)
Within the scope of the UI/UX Design Overhaul (Phase 8), Page Update [98] has been implemented. In this update, I made design enhancements to the GroupInvitation component on the CommunicationPage and implemented the functionality to list the group invitations for the user. Only authorized and authenticated users can view their own group invitations.

The name 'group requests' is not an accurate designation, so I organized it as 'group invitations'.