Page Update [100] (EventInvitation)
Within the scope of the UI/UX Design Overhaul (Phase 8), Page Update [100] has been implemented. In this update, I made design enhancements to the EventInvitation component on the CommunicationPage and implemented the functionality to list the event invitations for the user. Only authorized and authenticated users can view their own event invitations.

The term "event requests" is not an accurate designation, so I organized it as "event invitations".