Page Update [101] (FriendRequests)
Within the scope of the UI/UX Design Overhaul (Phase 8), Page Update [101] has been implemented. In this update, I made design enhancements to the FriendRequests component on the CommunicationPage and implemented the functionality to list the friend requests for the user. Only authorized and authenticated users can view their own friend requests.