Page Update [102] (Profile)
Within the scope of the UI/UX Design Overhaul (Phase 8), Page Update [102] has been implemented. In this update, I made design enhancements to the ProfileHeader and ProfileMainContent components on the Profile page. Additionally, I provided adjustments for the Add Friend and Follow buttons, ensuring that buttons for sending requests to oneself are no longer visible.