Part Update [75]
Within the scope of the UI/UX Design Overhaul (Phase 9), Part Update [75] has been implemented. In this update, I reorganized the group details and the list of groups the user is a member of within the ProfileGroupsContent component on the user's Profile page. I made necessary adjustments for profile examination and performed related modifications on the checkers.