Update Api & VoiceCallProvider
Within the scope of the UI/UX Design Overhaul (Phase 9), Update Api & VoiceCallProvider were implemented. As part of this update, I added functionality for CreateActivity to the API. In addition, I resolved an issue in VoiceCallProvider where rtcPeerConnectionRef was getting stuck and attempting to open a new connection repeatedly when frozen.