Part Update [85]
Within the scope of the UI/UX Design Overhaul (Phase 9), Part Update [85] has been implemented. In this update, I made adjustments to the display of followers and followed users in the ProfileHeader component. I also prepared the necessary settings for the new UI elements to be added and defined the required interfaces on the profile page.