Page Update [116]
Within the scope of the UI/UX Design Overhaul (Phase 9), Page Update [116] (PostDetails) has been implemented. In this update, I made design improvements to the PostDetails page and made significant changes to the API and interface definitions. Additionally, I transferred the functionality for liking posts and adding comments to the API.