Page Update [124]
Within the scope of the UI/UX Design Overhaul (Phase 9), Page Update [124] (MessagePage) was implemented. In this update, I made significant changes to the message chat data coming from the API on the MessagePage. Now, this page will receive updates concerning message encryption and the messaging process via the API, as well as updates on verifying message ownership.