Page Update [125] (Messaging Unlock Signature)
Within the scope of the UI/UX Design Overhaul (Phase 9) and Messaging Operations: Proof of messages with o1js, implementation of messaging encryption, and mutual signing for secure communication., Page Update [125] (MessagePage) has been implemented. In this update, on the MessagePage, I've written the necessary procedures to sign the timestamp with users' current sessions' AuthTokens, and send the signed data to the provider server on the API to unlock the messaging process.