Update Api & Page Update [126]
Within the UI/UX Design Overhaul (Phase 9), Page Update [126] (BlogDetail) has been implemented. In this update, I've added the feature for users to send a message request to the author on the BlogDetail page. Additionally, I've included the PostUnclockChatData method in the API for users to unlock chat locks through the API.