Page Update [135] & Update Api
Within the scope of the UI/UX Design Overhaul (Phase 9), Page Update [135] has been implemented. In this update, I connected the EventInvitation and message sending actions to the API on the Communication and Message pages. Utilizing Proto-Kit for messaging and operations and retrieving results from the Prover server through the MogartApi is currently the most stable approach. I am continuing to work on the necessary steps for moving this functionality to the client side.